Styling + Database
- Styling
 + main menu
 + register
 + login
- Database
 + adding new database on laptop
 + has a log in on myphpadmin
 + able to save data on database
