+++ mainmenu/Assets/Material/Color.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b76f4d0efe922444e8d79781fe8e28bc
+guid: 2a4dbc6db10296f4f869be795202c28c
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
