+++ mainmenu/Assets/Material/Color/background color.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f16974909160e4f4e8a92225cd479b25
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
