+++ mainmenu/Assets/Scenes/04 game menu.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7bbe90b10ac878e40974a0f37071fdce
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
