+++ mainmenu/Assets/Scenes/09 testing score.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0ce0417e0b562d045b674e3d18baf3f7
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
