+++ mainmenu/Assets/Script/DBManager.cs
@@ -7,9 +7,9 @@ public static class DBManager
 {
     public static string username;
     public static int score;
-    public static int duration;
-    public static int dbtest;
-    public static decimal dbfloat;
+    //public static int duration;
+    //public static int dbtest;
+    //public static decimal dbfloat;
 
     //use to check are we logged in 
     public static bool LoggedIn
