--- mainmenu/Assets/Script/Game.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2888c47690840ed4aaa8551ab7adbd35
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
