+++ mainmenu/Assets/XR.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: abefa33e9079f32469cfafd93f3fda95
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
