+++ mainmenu/Assets/XR/Loaders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4f93050ffb97c2040a7f8f958682a285
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
