+++ mainmenu/Assets/XR/Loaders/Open VR Loader.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b6691ca18b604b4cb2b468a0197322a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
