+++ mainmenu/Assets/XR/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a931a88ca3995ed41abca67cd2ae04bf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
