+++ mainmenu/Assets/XR/Settings/Open VR Settings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8365efcea0da0da4aa56e180b9e6fa30
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
