+++ mainmenu/Assets/XR/XRGeneralSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 67ee2acdec541a443be9daff4295afb6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
