+++ mainmenu/ProjectSettings/PackageManagerSettings.asset
@@ -9,9 +9,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 13964, guid: 0000000000000000e000000000000000, type: 0}
   m_Name: 
-  m_EditorClassIdentifier: UnityEditor:UnityEditor.PackageManager.UI:PackageManagerProjectSettings
+  m_EditorClassIdentifier: 
+  m_EnablePreviewPackages: 0
+  m_EnablePackageDependencies: 0
+  m_AdvancedSettingsExpanded: 1
   m_ScopedRegistriesSettingsExpanded: 1
   oneTimeWarningShown: 0
   m_Registries:
@@ -20,19 +23,32 @@ MonoBehaviour:
     m_Url: https://packages.unity.com
     m_Scopes: []
     m_IsDefault: 1
+    m_Capabilities: 7
+  - m_Id: scoped:Valve
+    m_Name: Valve
+    m_Url: https://registry.npmjs.org
+    m_Scopes:
+    - com.valvesoftware
+    - com.valvesoftware.unity.openvr
+    m_IsDefault: 0
+    m_Capabilities: 0
   m_UserSelectedRegistryName: 
   m_UserAddingNewScopedRegistry: 0
   m_RegistryInfoDraft:
     m_ErrorMessage: 
     m_Original:
-      m_Id: 
-      m_Name: 
-      m_Url: 
-      m_Scopes: []
+      m_Id: scoped:Valve
+      m_Name: Valve
+      m_Url: https://registry.npmjs.org
+      m_Scopes:
+      - com.valvesoftware
+      - com.valvesoftware.unity.openvr
       m_IsDefault: 0
+      m_Capabilities: 0
     m_Modified: 0
-    m_Name: 
-    m_Url: 
+    m_Name: Valve
+    m_Url: https://registry.npmjs.org
     m_Scopes:
-    - 
+    - com.valvesoftware
+    - com.valvesoftware.unity.openvr
     m_SelectedScopeIndex: 0
