+++ mainmenu/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.4.13f1
-m_EditorVersionWithRevision: 2019.4.13f1 (518737b1de84)
+m_EditorVersion: 2020.1.10f1
+m_EditorVersionWithRevision: 2020.1.10f1 (974a9d56f159)
