+++ mainmenu/ProjectSettings/XRPackageSettings.asset
@@ -1,0 +1,5 @@
+{
+    "m_Settings": [
+        "RemoveLegacyInputHelpersForReload"
+    ]
+}
